Update CompressApplication.php
This PR has mainly been opened as a 'talking point' and to work out the size limitations for Laravel Apps & their Vendor files on Vapor. As an application I'm developing has more than 45 MB of assets outside of the Vendor and Public directories.

When uploading direct to Lambda, the limit is 50MB, however when uploading to Lambda via S3, which from what I can tell Vapor is doing, the limit is 250 MB according to articles that have tested limits on Lambda. 

The Serverless CLI is able to achieve sizes up to 250MB when using S3 as the source.

https://dashbird.io/blog/exploring-lambda-limitations/

Removing this limit within my local copy of Vapor CLI allows me to upload a file that's double the current limit and appears to function perfectly fine?

https://www.dropbox.com/s/nco1adnqemx265v/100mb.PNG?dl=0

https://resilient-pond-csvcammlc6kk.vapor-farm-a1.com/

However there still appears to be something in the way of allowing larger files. It uploads successfully to S3, but during the 'Updating Function Configurations' stage it fails, looking at the Lambda configuration, there's no env variables etc set, it just appears to be a blank Lambda setup.

https://www.dropbox.com/s/8y9yt8g0b0y6azw/150mb-error.PNG?dl=0

Apologies if it's fired any alerts off your way!